Css commit
Added the background color For every input, select, text area element that is focus. Added the background of the fnBox, lnBox, addBox, mailBox, phoneBox, and idBox controls have the focus and are valid and added an valid checker logo on the right side of these fields and the rgb is green. Added the background of the fnBox, lnBox, addBox, mailBox, phoneBox, and idBox controls have the focus and are invalid and added an invalid checker logo on the right side of these fields and the rgb is red